last one

+++ Utilities/DAO/DAOImplement/EmployeeDAO.cs
@@ -107,68 +107,77 @@ namespace MrBricolage.Utilities.DAO.DAOImplement
         {
             bool flag = true;
             
-            try
+            if (empTocreatee != null)
             {
-               
-
-
-
-                string sql = string.Format("SELECT emp_name , emp_f_name FROM employee WHERE emp_name = '{0}' AND emp_f_name = '{1}' ; SELECT Login  FROM employee WHERE Login = '{2}' ;", empTocreatee.Name , empTocreatee.F_Name, empTocreatee.Login);
-
-
-
-                MySqlCommand cmd = new MySqlCommand(sql, conn);
-                MySqlDataReader reader = cmd.ExecuteReader();
-
-                if (reader.Read())
+                try
                 {
 
-                    MessageBox.Show("l'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " existe deja dans votre DB", "Infos");
-                    flag = false;
-                    
-                }else
-                {
-                    reader.NextResult();
+
+
+
+                    string sql = string.Format("SELECT emp_name , emp_f_name FROM employee WHERE emp_name = '{0}' AND emp_f_name = '{1}' ; SELECT Login  FROM employee WHERE Login = '{2}' ;", empTocreatee.Name, empTocreatee.F_Name, empTocreatee.Login);
+
+
+
+                    MySqlCommand cmd = new MySqlCommand(sql, conn);
+                    MySqlDataReader reader = cmd.ExecuteReader();
+
                     if (reader.Read())
                     {
-                        MessageBox.Show("Login existe deja dans la DB , changé votre Login SVP !", "Infos");
+
+                        MessageBox.Show("l'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " existe deja dans votre DB", "Infos");
                         flag = false;
-                        
+
                     }
                     else
                     {
-                        reader.Close();
-                        sql = string.Format("INSERT INTO employee (emp_name , emp_f_name , Login , _password) VALUES ('{0}','{1}','{2}','{3}');", empTocreatee.Name, empTocreatee.F_Name, empTocreatee.Login, empTocreatee.Password);
+                        reader.NextResult();
+                        if (reader.Read())
+                        {
+                            MessageBox.Show("Login existe deja dans la DB , changé votre Login SVP !", "Infos");
+                            flag = false;
 
-                       
+                        }
+                        else
+                        {
+                            reader.Close();
+                            sql = string.Format("INSERT INTO employee (emp_name , emp_f_name , Login , _password) VALUES ('{0}','{1}','{2}','{3}');", empTocreatee.Name, empTocreatee.F_Name, empTocreatee.Login, empTocreatee.Password);
 
-                        MySqlCommand cmd2 = new MySqlCommand(sql, conn);
-                        MySqlDataReader reader2 = cmd2.ExecuteReader();
 
 
-                        MessageBox.Show("L'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " a bien ete insere dans votre DB ! ", "Infos");
-                        reader2.Close();
+                            MySqlCommand cmd2 = new MySqlCommand(sql, conn);
+                            MySqlDataReader reader2 = cmd2.ExecuteReader();
+
+
+                            MessageBox.Show("L'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " a bien ete insere dans votre DB ! ", "Infos");
+                            reader2.Close();
+                        }//end if 
+
                     }//end if 
 
-                }//end if 
 
 
 
 
-
-                reader.Close();
-
+                    reader.Close();
 
 
 
-            } catch(Exception ex)
+
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    MessageBox.Show("problem d'insertion d'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " dans votre DB", "Infos");
+                    flag = false;
+
+
+                }//end try catch
+            }else
             {
-                Console.WriteLine(ex.Message);
-                MessageBox.Show("problem d'insertion d'employee " + empTocreatee.Name + " " + empTocreatee.F_Name + " dans votre DB", "Infos");
                 flag = false;
-
-
-            }//end try catch
+                MessageBox.Show("L'employee est null !", "Infos");
+            }
 
             return flag;
            
@@ -183,24 +192,31 @@ namespace MrBricolage.Utilities.DAO.DAOImplement
         {
             bool flag = true;
 
-            try
+           if (EmployeeToDelete != null)
             {
+                try
+                {
 
-                string sql = string.Format("DELETE FROM list_of_art WHERE id_facture in (SELECTE ID_f FROM facture WHERE emp_num = {0} ; ); DELETE FROM facture WHERE emp_nom = {0} ; DELETE FROM  employee WHERE id_emp = {0};", EmployeeToDelete.Id);
+                    string sql = string.Format("DELETE FROM list_of_art WHERE id_facture in (SELECTE ID_f FROM facture WHERE emp_num = {0} ; ); DELETE FROM facture WHERE emp_nom = {0} ; DELETE FROM  employee WHERE id_emp = {0};", EmployeeToDelete.Id);
 
 
-                MySqlCommand cmd = new MySqlCommand(sql, conn);
-                MySqlDataReader reader = cmd.ExecuteReader();
+                    MySqlCommand cmd = new MySqlCommand(sql, conn);
+                    MySqlDataReader reader = cmd.ExecuteReader();
 
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    flag = false;
+                }//end trycatch
+            }else
+            {
+                flag = false;
+                MessageBox.Show("L'employee est null !", "Infos");
             }
-            catch(Exception ex)
-            {
-                Console.WriteLine(ex.Message);
-                flag= false;
-            }//end trycatch
 
             return flag;
-        }
+        }//end delete
 
 
         public override bool update(Employee EmployeeToUpdate) 
@@ -209,23 +225,26 @@ namespace MrBricolage.Utilities.DAO.DAOImplement
             bool flag = true;
 
 
-            try
+            if (EmployeeToUpdate!= null)
             {
-                string sql =string.Format("UPDATE employee SET emp_name = '{0}' , emp_f_name = '{1}' , Login = '{2}' , _password = '{3}' WHERE id_emp = {4} ", EmployeeToUpdate.Name, EmployeeToUpdate.F_Name, EmployeeToUpdate.Login, EmployeeToUpdate.Password, EmployeeToUpdate.Id);
+                try
+                {
+                    string sql = string.Format("UPDATE employee SET emp_name = '{0}' , emp_f_name = '{1}' , Login = '{2}' , _password = '{3}' WHERE id_emp = {4} ", EmployeeToUpdate.Name, EmployeeToUpdate.F_Name, EmployeeToUpdate.Login, EmployeeToUpdate.Password, EmployeeToUpdate.Id);
 
 
-                MySqlCommand cmd = new MySqlCommand(sql, conn);
-                MySqlDataReader reader = cmd.ExecuteReader();
+                    MySqlCommand cmd = new MySqlCommand(sql, conn);
+                    MySqlDataReader reader = cmd.ExecuteReader();
 
-
-
-
-
-            }
-            catch(Exception ex)
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    flag = false;
+                }//end trycatch 
+            }else
             {
-                Console.WriteLine(ex.Message);
                 flag = false;
+                MessageBox.Show("L'employee est null !", "Infos");
             }
 
             return flag;
